В AdminCommentController метод deleteAdmin метод deleteAdmin переименован в deleteCommentById. Методы findById, после которых не требовалось работать с найденной сущностью, заменены на existsById

+++ main/src/main/java/ru/practicum/service/AdminCommentService.java
@@ -14,8 +14,11 @@ public class AdminCommentService {
     private final CommentRepository commentRepository;
 
     public void deleteCommentById(Long commentId) {
-        commentRepository.findById(commentId).orElseThrow(() -> new CommentNotFoundException(commentId.toString()));
-        commentRepository.deleteById(commentId);
+        if (commentRepository.existsById(commentId)) {
+            commentRepository.deleteById(commentId);
+        } else {
+            throw new CommentNotFoundException(commentId.toString());
+        }
     }
 
 }
